Add users table and make table component reusable
